Add dynamicly generated locale-specific accessors of the form 'content_en_us'.  Two purposes:
This allows efficient editing of translations in a straightfoward and robust way.  The Rails' nested_attributes methods some people have built suffer from the fact that Rails' nested attributes require the translation id to identify an existing translation, which can lead to failed updates or duplicates even though a translation is uniquely identifiable by locale + foreign_key.  The existing set_translations method is handy, but it doesn't map well to forms and seems potentially less effective then this approach.

A nice thing about the dynamic readers is that they bypass I18n fallbacks so that you can unambiguously check for exact locales, which is what you want in editing forms most of the time.

This patch requires I18n.available_locales to be defined before the translates method is called on any models.