Guarantee that translations are always updated when using the dynamic locale-specific writers while still maintaining an efficient implementation.
Adapter#save_translations! now calls Adapter#build_translations which has an option either to save or not.  This allows the dynamic writers to ensure that the actual translations are updated without necessarily touching the DB (ie. you don't want a single attribute writer to trigger a full save of all translations ever time...).  Note that the implementation of build_translations is more or less the same as save_translations was before, but the use of record.translations.find_or_initialize_by_locale was replaced with `detect || build` logic instead because that ensures that the same physical ruby object is updated, which find_or_initialize_by_locale does not provide that benefit.