Fix heisenbug that was preventing is_blank translations from being deleted in production.
I could not produce a failing test yet, and I just need to push this out to production.  What happened was that depending on the object I submitted, I would get some errors from blank translations that should have been deleted.  It wouldn't give me errors on ALL translations, just some of them, and it was dependent on the specific base object I was working, however it stayed consistent within that object.  I suspect some hash ordering effect is at play.  I have a stash of an attempted test case, but no time to work on it right at the moment.